Extract testing enum for reuse
